+++ vmpop3d/entry.sh
@@ -1,10 +1,8 @@
 #!/bin/sh
 
-touch /var/run/rsyslogd.pid
-rm /var/run/rsyslogd.pid
+if [ -f /var/run/rsyslogd.pid ]; then rm /var/run/rsyslogd.pid; fi
 /usr/sbin/rsyslogd -n &
 
 vm-pop3d -d --user vmail
 
-touch /var/log/maillog
 tail -f /var/log/maillog
